pyproject.toml
[tool.poetry]
name = "testenv"
version = "0.1.0"
description = ""
authors = ["Stepan <ssprudnik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^1.10.9"
fastapi = "^0.97.0"
sqlalchemy = "^2.0.16"
uvicorn = "^0.22.0"
asyncpg = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
